+++ kite-core/src/main/kotlin/com/tang/kite/sql/dialect/SqlDialect.kt
@@ -14,6 +14,8 @@ interface SqlDialect {
 
     fun supportsCascade(): Boolean
 
+    fun supportsCommentOnTable(): Boolean
+
     fun supportsCommentOnColumn(): Boolean
 
     fun requiresTableForDropIndex(): Boolean